[Inductor cutlass backend] Fix cutlass_utils.get_max_alignment() for strided layouts.
Fixes cutlass_utils.get_max_alignment() which was so far not checking the alignment properly. Basically
the method so far assumed that the passed layout is contiguous and row-major, which does not have to be true.

Test Plan:
CI - test_cutlass_backend.py to prevent regressions
TODO: Add unit test

ghstack-source-id: 51947c5a8acb4e4cc1c57487b0aa4529833078b1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124930